fix(conformance): address sdk-review findings on C001/C002 remediation
- catch UnicodeDecodeError alongside OSError in bootstrap's file-read helpers
- report touched_files as repo-relative paths, matching its documented contract
- split bootstrap/command.py into autodetect.py + args.py, hoist redundant
  render imports to module level
- give the C002 checker's exit-zero extraction the same renovate.json
  fallback as bootstrap's own --enforce autodetection
- add a -- separator before <ref> in C001's git ls-remote fallback
- extend the orthogonal-gate skip branch to compile-check touched .py files
- add a structural forces_external_influence rule field so C001's residue
  escalation doesn't depend solely on the model setting it per invocation
- document C003's split autofixable semantics (absent-file is mechanical,
  missing-entry is judgment-only)
- add tests for the malformed pyproject.toml guard case and the exit-zero
  fallback; regenerate stale rule-doc output
